Create README.md
This script is part of a Chrome extension that displays random React/JavaScript interview questions and answers from a local JSON file. When the extension's popup is opened, it fetches the questions and answers from the JSON file. Each time the "Next Question" button is clicked, a random question and its corresponding answer are displayed. The script uses the fetch API to retrieve the JSON data and the Math.random() function to randomly select a question from the list.